requirements.txt
Django==3.0.8
asgiref==3.2.10
django-rest-framework==0.1.0
djangorestframework==3.11.0
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1